changing hls functions to simple precision

+++ src/stm.cpp
@@ -44,11 +44,11 @@ void stm(int enable_stm, int num_pixels, hls::stream<pixel_out_t> &in, hls::stre
 
       if(enable_stm) {
     	  // softplus
-    	  v_soft = hls::logf(hls::exp(v_in) + 1); //logf for single-precision natural logarithm
+    	  v_soft = hls::logf(hls::expf(v_in) + 1); //logf for single-precision natural logarithm
                                                 //log for double-precision natural logarithm
 
     	  //tanh
-    	  v_tanh = hls::tanh(v_soft);
+    	  v_tanh = hls::tanhf(v_soft);
 
     	  //mult
     	  v_mult = v_tanh * v_in;
